chore: #37 VisitedState 파일 분리
