+++ Projects/Feature/TrashDetail/TrashDetailFeature/Sources/Features/VisitedState.swift
@@ -1,0 +1,27 @@
+//
+//  VisitedState.swift
+//  TrashDetailFeature
+//
+//  Created by Jiyeon on 7/13/25.
+//  Copyright © 2025 Sseudam.a2bo.ios. All rights reserved.
+//
+
+import Foundation
+
+public enum VisitedState {
+  /// 방문 인증 5분 지나지 않음
+  case remainTime
+  /// 쓰레기통이 멀리(5m 밖)에 있음
+  case far
+  /// 5m 안 쪽에 있음
+  case enableVisit
+  
+  case unknown
+  
+  case denyPermission
+  
+  var buttonEnable: Bool {
+    self == .enableVisit
+  }
+  
+}
